# docs/api/openapi.json
{"openapi":"3.0.0","paths":{"/v1/health":{"get":{"operationId":"HealthController_health","parameters":[],"responses":{"200":{"description":""}},"tags":["health"]}},"/v1/ready":{"get":{"operationId":"HealthController_ready","parameters":[],"responses":{"200":{"description":""}},"tags":["health"]}},"/v1/me":{"get":{"operationId":"MeController_getMe","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["me"]},"patch":{"operationId":"MeController_update","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["me"]},"delete":{"operationId":"MeController_requestDeletion","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["me"]}},"/v1/me/export":{"get":{"operationId":"MeController_exportData","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["me"]}},"/v1/me/devices":{"post":{"operationId":"MeController_registerDevice","parameters":[],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["me"]}},"/v1/characters":{"get":{"operationId":"CharactersController_list","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["characters"]},"post":{"operationId":"CharactersController_create","parameters":[],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["characters"]}},"/v1/characters/{id}":{"get":{"operationId":"CharactersController_get","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["characters"]},"delete":{"operationId":"CharactersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["characters"]}},"/v1/uploads":{"post":{"operationId":"UploadsController_create","parameters":[],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["uploads"]}},"/v1/uploads/{id}/complete":{"post":{"operationId":"UploadsController_complete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["uploads"]}},"/v1/comics":{"get":{"operationId":"ComicsController_list","parameters":[{"name":"limit","required":true,"in":"query","schema":{"type":"string"}},{"name":"cursor","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["comics"]},"post":{"operationId":"ComicsController_create","parameters":[],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["comics"]}},"/v1/comics/{id}":{"get":{"operationId":"ComicsController_get","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["comics"]}},"/v1/comics/{id}/status":{"get":{"operationId":"ComicsController_status","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["comics"]}},"/v1/comics/{id}/share":{"post":{"operationId":"ComicsController_share","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["comics"]}},"/v1/feed":{"get":{"operationId":"SocialController_feed","parameters":[{"name":"limit","required":true,"in":"query","schema":{"type":"string"}},{"name":"cursor","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["social"]}},"/v1/discover":{"get":{"operationId":"SocialController_discover","parameters":[{"name":"limit","required":true,"in":"query","schema":{"type":"string"}},{"name":"cursor","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["social"]}},"/v1/reels":{"get":{"operationId":"SocialController_reels","parameters":[{"name":"limit","required":true,"in":"query","schema":{"type":"string"}},{"name":"cursor","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["social"]}},"/v1/comics/{id}/like":{"post":{"operationId":"SocialController_like","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["social"]},"delete":{"operationId":"SocialController_unlike","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["social"]}},"/v1/comics/{id}/comments":{"get":{"operationId":"SocialController_listComments","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"limit","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["social"]},"post":{"operationId":"SocialController_addComment","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["social"]}},"/v1/comments/{id}":{"delete":{"operationId":"SocialController_deleteComment","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["social"]}},"/v1/comics/{id}/repost":{"post":{"operationId":"SocialController_repost","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["social"]}},"/v1/comics/{id}/view":{"post":{"operationId":"SocialController_view","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["social"]}},"/v1/users/{id}/follow":{"post":{"operationId":"SocialController_follow","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["social"]},"delete":{"operationId":"SocialController_unfollow","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["social"]}},"/v1/public/comics/{slug}":{"get":{"operationId":"PublicController_getBySlug","parameters":[{"name":"slug","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["public"]}},"/v1/admin/users":{"get":{"operationId":"AdminController_users","parameters":[{"name":"limit","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["admin"]}},"/v1/admin/comics":{"get":{"operationId":"AdminController_moderationQueue","parameters":[{"name":"limit","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["admin"]}},"/v1/admin/comics/{id}/moderate":{"post":{"operationId":"AdminController_moderate","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["admin"]}},"/v1/admin/users/{id}/credits":{"post":{"operationId":"AdminController_grant","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["admin"]}},"/v1/admin/metrics":{"get":{"operationId":"AdminController_metrics","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["admin"]}}},"info":{"title":"StoryMe API","description":"Backend API for StoryMe — personalized comic stories + social feed.","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{}}}